% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_ttest.R
\name{bayesian_ttest}
\alias{bayesian_ttest}
\title{Bayesian t-Test for UX Comparisons}
\usage{
bayesian_ttest(group1, group2, n_samples = 10000)
}
\arguments{
\item{group1}{A numeric vector representing values from the first group.}

\item{group2}{A numeric vector representing values from the second group.}

\item{n_samples}{The number of samples to draw from the posterior distribution (default is 10000).}
}
\value{
The probability that the mean of group1 is greater than group2.
}
\description{
Performs a Bayesian t-test to estimate the probability that two groups have different means.
}
\examples{
set.seed(42)
group1 <- rnorm(30, mean = 50, sd = 10)
group2 <- rnorm(30, mean = 45, sd = 10)
bayesian_ttest(group1, group2)
}
