% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_mean_ci}
\alias{bootstrap_mean_ci}
\title{Get a confidence interval for the mean using bootstrapping}
\usage{
bootstrap_mean_ci(data, n_boot = 1000)
}
\arguments{
\item{data}{A numeric vector}

\item{n_boot}{Number of bootstrap samples}
}
\value{
Bootstrap confidence interval
}
\description{
Get a confidence interval for the mean using bootstrapping
}
