% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_test.R
\name{chi_square_test}
\alias{chi_square_test}
\title{Chi-Square Test for Categorical Data}
\usage{
chi_square_test(data, categorical_col1, categorical_col2)
}
\arguments{
\item{data}{A data frame containing UX categorical data.}

\item{categorical_col1}{The first categorical column name.}

\item{categorical_col2}{The second categorical column name.}
}
\value{
A list with Chi-Square test results.
}
\description{
Performs a chi-square test to determine if there is a significant association between two categorical variables.
}
\examples{
test_data <- data.frame(UI_Version = rep(c("A", "B"), each = 50),
                        Task_Success = sample(c("Success", "Fail"), 100, replace = TRUE))
chi_square_test(test_data, "UI_Version", "Task_Success")
}
