% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Get quick descriptive stats on your dataset}
\usage{
descriptive_stats(data, cols)
}
\arguments{
\item{data}{A data frame}

\item{cols}{Vector of column names}
}
\value{
A summary table with mean, median, SD, and range
}
\description{
Get quick descriptive stats on your dataset
}
