% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UXtoolbox-package.R
\docType{package}
\name{UXtoolbox-package}
\alias{UXtoolbox}
\alias{UXtoolbox-package}
\title{UXtoolbox: A Comprehensive Toolkit for UX and Human Factors Research}
\description{
The UXtoolbox package provides tools for analyzing user experience (UX) data, including Bayesian statistical tests, predictive modeling, usability heuristics, and cognitive modeling.
}
\details{
\subsection{Features}{
\itemize{
\item Bayesian ANOVA, t-tests, regression, and survival analysis
\item Mixed-effects and structural equation modeling (SEM)
\item Implementation of UX laws such as Fitts' Law, Hick's Law, and Weber's Law
\item Content and sentiment analysis tools
\item Visualization tools for UX data
}
}

\subsection{Installation}{

To install the package from GitHub:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{install.packages("devtools")  # Install devtools if not already installed
devtools::install_github("mohsen-rafiei/UXtoolbox")
}\if{html}{\out{</div>}}
}

\subsection{Getting Started}{

Load the package:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(UXtoolbox)
}\if{html}{\out{</div>}}

Perform Bayesian ANOVA:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{test_data <- data.frame(
  completion_time = c(5.1, 6.3, 5.5, 4.8, 7.2, 6.0, 5.4, 4.9, 7.0, 6.5),
  UI_version = factor(rep(c("A", "B", "C"), length.out = 10))
)
bayesian_anova(test_data, completion_time ~ UI_version)
}\if{html}{\out{</div>}}
}

\subsection{Package Documentation}{

To view a list of functions in the package, run:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{help(package = "UXtoolbox")
}\if{html}{\out{</div>}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mohsen-rafiei/UXtoolbox}
  \item Report bugs at \url{https://github.com/mohsen-rafiei/UXtoolbox/issues}
}

}
\author{
\strong{Maintainer}: Mohsen Rafiei \email{moo.rafiei@gmail.com} (\href{https://orcid.org/0000-0002-5136-1733}{ORCID})

}
