% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tests.R
\name{compare_groups_ttest}
\alias{compare_groups_ttest}
\title{Compare two groups with a t-test}
\usage{
compare_groups_ttest(data, group_col, value_col)
}
\arguments{
\item{data}{A data frame}

\item{group_col}{Column name for grouping variable}

\item{value_col}{Column name for numeric values}
}
\value{
A t-test result
}
\description{
Compare two groups with a t-test
}
